feat: fetced the data from the firebase and show in the card
