+++ src/components/petCard/PetCard.tsx
@@ -48,6 +48,9 @@ const PetCard: React.FC<PetCardProps> = ({
   );
 };
 
+
+
+export default PetCard;
 const styles = StyleSheet.create({
   container: {
     flexDirection: 'row',
@@ -132,5 +135,3 @@ const styles = StyleSheet.create({
     alignItems: 'center',
   },
 });
-
-export default PetCard;
